v55-fix: 修复 SAGE 训练 CUDA/CUBLAS 稳定性与 OOM 问题；验证 val_auc=0.8814 val_aupr=0.7870

+++ L4/src/iron_aging_gnn/models/sage.py
@@ -187,12 +187,12 @@ class SAGELinkPredictor(nn.Module):
             raise ValueError(
                 f"蛋白输入维度 {prot_x.shape[-1]} < prot_esm_dim {self.prot_esm_dim}"
             )
-        comp_x_actual = comp_x[:, :self.comp_feat_dim]
-        prot_esm = prot_x[:, :self.prot_esm_dim]
+        comp_x_actual = comp_x[:, :self.comp_feat_dim].float()
+        prot_esm = prot_x[:, :self.prot_esm_dim].float()
         prot_h = self.prot_feat_proj(prot_esm)
 
         if use_pathway and self.pathway_proj is not None and self.n_pathways > 0:
-            prot_pathway = prot_x[:, self.prot_esm_dim:self.prot_esm_dim + self.n_pathways]
+            prot_pathway = prot_x[:, self.prot_esm_dim:self.prot_esm_dim + self.n_pathways].float()
             prot_h = prot_h + self.pathway_proj(prot_pathway)
 
         comp_h = self.comp_proj(comp_x_actual)
